Ok, I have filled in all of the level, so the structure should be about finished.
Also, I added salt to the level.  The pit and the pit ramp are now covered and salt, and there is a salt can on platform3.  There will eventually be salt streaming from the can to explain the salt on the ground.

If Helix touches the salt, he should get hurt, obviously.

I think I did some other stuff, but I don't remember what.

git-svn-id: svn+ssh://leetcode.net/home/jreese/backup/srv/svn/repos/rit/2dgames/trunk@604 ad4848e8-a12d-0410-9a8f-f544a64f585b