Ok, I've merely updated some aspects to my level.
Maybe someone can figure a more clever way to do this, or maybe I'll do it later...

To use my level, change the name of Scenery to something else and change PatScenery to Scenery.

Then change Garden.lua to something else and change PatGarden to Garden.

Finally, clean the solution, and then rebuild.

Do this in reverse to switch back.  I'm going to bed.

git-svn-id: svn+ssh://leetcode.net/home/jreese/backup/srv/svn/repos/rit/2dgames/trunk@687 ad4848e8-a12d-0410-9a8f-f544a64f585b