Added Boost Powerup. Finished death animations.
Joe, you might want to tinker with the boost code in Helix.cs.  It's in the acceleration section under "if(flying)".  The boost is as little jerky.  The boost period ends, and Helix slows down abruptly instead of smootly.  I'm not sure how to do this because I'm not as familiar with the acceleration-gravity rules.

git-svn-id: svn+ssh://leetcode.net/home/jreese/backup/srv/svn/repos/rit/2dgames/trunk@763 ad4848e8-a12d-0410-9a8f-f544a64f585b